+++ src/App.js
@@ -1,10 +1,10 @@
 import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
-import Home from './components/CovidHome/Home';
 import './App.css';
 import { fetchCovidCases } from './redux/covid/covid';
 import Headers from './components/headers/headers';
-import CovidCases from './components/covidCaseHome/covidCase';
+import CovidCases from './components/HomePage/Category';
+import HeroSection from './components/HeroSection/HeroSection';
 
 function App() {
   const dispatch = useDispatch();
@@ -15,7 +15,7 @@ function App() {
   return (
     <div className="App">
       <Headers />
-      <Home />
+      <HeroSection />
       <CovidCases />
     </div>
   );
