+++ src/components/DetailsHeroSection/DetailsHeroSection.module.css
