+++ src/components/DetailsHome/detailsCategory.module.css
