--- src/components/covidCaseHome/covidCases.module.css
@@ -1,158 +1,0 @@
-* {
-  padding: 0;
-  margin: 0;
-}
-
-.container {
-  display: grid;
-  padding: 3em;
-  grid-template-columns: 1fr 1fr 1fr;
-}
-
-.covidHome {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  position: relative;
-  padding: 2em;
-  border: 0.04em solid rgba(139, 198, 247, 0.048);
-}
-
-.covidHome:nth-child(1) {
-  z-index: 100;
-  background-color: #446ab1;
-}
-
-.covidHome:nth-child(2) {
-  background-color: rgba(49, 76, 126, 0.534);
-}
-
-.covidHome:nth-child(4) {
-  z-index: 100;
-  background-color: #446ab1;
-}
-
-.covidHome:nth-child(3) {
-  background-color: rgba(49, 76, 126, 0.534);
-}
-
-.covidHome:nth-child(5) {
-  z-index: 100;
-  background-color: #446ab1;
-}
-
-.covidHome:nth-child(6) {
-  background-color: rgba(49, 76, 126, 0.534);
-}
-
-.covidHome:nth-child(8) {
-  z-index: 100;
-  background-color: #446ab1;
-}
-
-.covidHome:nth-child(7) {
-  background-color: rgba(49, 76, 126, 0.534);
-}
-
-.covidHome:nth-child(9) {
-  z-index: 100;
-  background-color: #446ab1;
-}
-
-.covidHome:nth-child(10) {
-  background-color: rgba(49, 76, 126, 0.534);
-}
-
-.img {
-  width: 8em;
-  height: 8em;
-}
-
-.font {
-  font-size: 1.5em;
-  position: absolute;
-  top: 1em;
-  right: 1em;
-  color: rgba(255, 253, 251, 0.671);
-}
-
-.heading h3 {
-  color: rgba(255, 255, 255, 0.877);
-  font-size: 2em;
-  padding: 0.2em 5em;
-  font-family: 'Roboto', sans-serif;
-  font-weight: 500;
-}
-
-.container h2 {
-  color: rgba(255, 253, 251, 0.815);
-  font-family: 'Poppins', sans-serif;
-  font-weight: 500;
-  margin-top: 1em;
-  font-size: 1.5em;
-}
-
-.loader {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100vh;
-  position: fixed;
-  background: radial-gradient(ellipse at center, #4292d4 0%, #3b7eb4 50%, #2e70a7 100%);
-  z-index: 100;
-  top: 0;
-}
-
-.loader img {
-  width: 80px;
-  height: 80px;
-  object-fit: cover;
-}
-
-@media (max-width: 790px) {
-  .container {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    justify-content: center;
-  }
-
-  .div {
-    display: flex;
-  }
-
-  .img {
-    width: 4em;
-    height: 4em;
-  }
-
-  .font {
-    padding-left: 2.6em;
-    padding-bottom: 3em;
-    padding-top: 0.2em;
-    font-size: 1.3em;
-  }
-
-  .heading h3 {
-    font-size: 1.3em;
-    padding: 0.1em 1.5em;
-  }
-
-  h2 {
-    font-size: 1.3em;
-    align-items: flex-end;
-  }
-
-  .loader img {
-    width: 50px;
-    height: 50px;
-  }
-}
-
-@media (max-width: 390px) {
-  .container {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-  }
-}
